refactor: extract shared AlcPathResolver for test alc-path lookups
DuplicatePackageTests and AutoStubAutoIncrementTests both resolved the
BC AL compiler binary via near-identical inline helpers. Move the logic
into AlcPathResolver and relocate the four ResolveAlcPath_* unit tests
into AlcPathResolverTests so coverage stays where the implementation lives.

OSX continues to map to linux/alc as before; a Mac-specific lookup is
deliberately not added until a Mac contributor can verify it.
